Move fpca to a dimensionality reduction module
