+++ skfda/exploratory/dim_reduction/projection/__init__.py
@@ -1,0 +1,1 @@
+from ._fpca import fpca
